refactor(agent): isolate reasoning replay recovery
Problem:\nThe reasoning replay fix exceeded repository function-size ratchets in streamWithFrozen and streamWithSamplingRecovery, and several test files crossed their per-file budgets.\n\nFix:\nExtract bounded reasoning and retry-finalization helpers, move provider capabilities into a focused file, and split replay tests by concern without changing request serialization or recovery behavior.\n\nVerification:\ngo run ./tools/repolint\ngolangci-lint run --timeout=5m ./...\ngo test ./...\ngo test -race ./internal/agent ./internal/control\ncd desktop && go test ./...\ngo vet ./...\n./scripts/cache-guard.sh
